remove market and dependency dialog provider
